chore(project) Refactor base directory and change JS mutant
