+++ .npmignore
@@ -1,2 +1,2 @@
 **
-!shared/target/js-2.13/weapon-regex-opt.js
+!core/target/js-2.13/weapon-regex-opt.js
