+++ core/src/main/scala/weaponregex/extension/StringExtension.scala
