+++ core/src/main/scala/weaponregex/model/Location.scala
