+++ core/src/main/scala/weaponregex/model/Position.scala
