+++ core/src/main/scala/weaponregex/model/mutation/TokenMutator.scala
