+++ core/src/main/scala/weaponregex/model/regextree/Leaf.scala
