+++ core/src/main/scala/weaponregex/model/regextree/Node.scala
