+++ core/src/main/scala/weaponregex/model/regextree/RegexTree.scala
