+++ core/src/main/scala/weaponregex/parser/Parser.scala
