+++ core/src/main/scalajs/weaponregex/WeaponRegeX.scala
