+++ core/src/main/scalajvm/weaponregex/WeaponRegeX.scala
