+++ core/src/test/scala/weaponregex/model/regextree/NodeTest.scala
