+++ core/src/test/scala/weaponregex/parser/ParserTest.scala
